The Bruce version ..
moved configuration to the pai_config.sh file. It is still relatively fragile; but should function adequately.

future .. more documentation, cleaner script code, ability to include a folder full of doc files (architecture guidelines).

pie in the sky: ability to point the AI at that reference docs folder via the AI so that we do not have to copy/paste it all in. It is possible ... but I am busy.